+++ frontend/src/pages/GuideDetailPage.css
@@ -394,9 +394,22 @@
   font-size: 0.88rem;
 }
 
-.gdp-match-login a {
+.gdp-match-login a,
+.gdp-match-login .gdp-match-login-btn {
   color: #047857;
   font-weight: 700;
+}
+
+.gdp-match-login-btn {
+  display: inline;
+  margin: 0;
+  padding: 0;
+  border: none;
+  background: none;
+  font: inherit;
+  cursor: pointer;
+  text-decoration: underline;
+  text-underline-offset: 2px;
 }
 
 .gdp-match-guide-warn {
